refactor(core): rename AgentMessageEvent → ChatOutputEvent, polish slash commands
Address review comments:
- Move AgentMessageEvent from agent/entity.rs to primitives/ as
  ChatOutputEvent — it's a shared primitive, not agent-specific
- Rename Done variant to AssistantDone
- Add SlashCommandOutput variant for slash command responses
- Inline registry.rs into slash_command/mod.rs (remove extra file)
- Make SlashCommand::execute and SlashCommands::process async
- Slash commands now emit SlashCommandOutput instead of AssistantText
- Update SSE event name mapping in web layer
