+++ core/src/slash_command/commands/ping.rs
@@ -3,6 +3,7 @@ use super::super::traits::{SlashCommand, SlashCommandContext, SlashCommandOutput
 
 pub struct PingCommand;
 
+#[async_trait::async_trait]
 impl SlashCommand for PingCommand {
     fn name(&self) -> &str {
         "ping"
@@ -12,7 +13,7 @@ impl SlashCommand for PingCommand {
         "Responds with pong"
     }
 
-    fn execute(
+    async fn execute(
         &self,
         _ctx: &SlashCommandContext,
         _args: &str,
